chore(*) remove dataplane reconciller, change discovery plugin
